SCRUM-209 refactor: move RegisteredUsers from widgets to entities, change RegisteredUsers logic
