+++ src/entities/Public/ui/RegisteredUsers/RegisteredUsers.module.scss
