+++ src/entities/Public/ui/index.ts
@@ -1,0 +1,1 @@
+export { RegisteredUsers } from './RegisteredUsers/RegisteredUsers'
